requirements.txt
Flask>=2.0
python-telegram-bot==13.15
prettytable